--[ The Crafting Update ]--
MMOItem crafting is now handled vanilla style.
Recipe types supported are: Shaped, Shapeless,
Furnace, Blast Furnace and Smoker

Additional Changes:
- The Advanced Workbench has been removed
- Custom NBT Tags can now be applied to items
- Fixed Elements menu back button
- Changed a long lore line in Custom Sounds GUI
- Items with 'disable interactions' can no longer be thrown

Important: This is a test version and bugs are to be expected. A few QoL features will be added soon and currently it is only supported on version 1.14 and later.