Added cancellable events for item damage.
New stat 'Amphibian' which makes items only work underwater, or under lava, or such.

Currently, this stat and the Biome Requirement only actually prevent you from equipping items in other places, but equipping the item and not changing your laodout allows you to use them elsewhere... I'll be working on a more dynamic check of requirements. This is not a problem for Gem Stones and Consumables though :o