Updating MMOItems with Random Stats now asks the following question:
Under the new RNG parameters, is the chance of getting this same RNG roll less than 1 in 3,000?

If the answer is *yes*, the stat is Rerolled with the new parameters. Otherwise it remains the same :)

Or then again, you can always enable in the config, if you wanted them to roll
```yml
item-revision:
    reroll-when-updated: true
```