- Added `MATH(equation)` to the stats.yml file.
I got this idea because of the Movement Speed system. The default Minecraft Movement Speed is 0.1, but IMO having 0.1 appear on Items, doesn't seem like a significant increase, even though you're actually doubling the Player's speed.

So now you can add to your stats.yml `movement-speed: 'Speed: +MATH(#*10)%'`
Now if I set the movement speed stat on an item to 0.1, the item will display in its lore "Speed: +100%"

This is purely "cosmetic" meaning this will only affect how the specific Item Stat is displayed in the Lore, not the stat itself!